+++ src/icalendar/tests/prop/test_unit.py
@@ -44,6 +44,9 @@ class TestProp(unittest.TestCase):
 
         dt_list = vDDDLists([datetime(2000, 1, 1), datetime(2000, 11, 11)])
         self.assertEqual(dt_list.to_ical(), b'20000101T000000,20001111T000000')
+        
+        instance = vDDDLists([])
+        self.assertFalse(instance == "value")
 
     def test_prop_vDate(self):
         from icalendar.prop import vDate
@@ -110,6 +113,7 @@ class TestProp(unittest.TestCase):
         self.assertRaises(ValueError, vFrequency, 'bad test')
         self.assertEqual(vFrequency('daily').to_ical(), b'DAILY')
         self.assertEqual(vFrequency('daily').from_ical('MONTHLY'), 'MONTHLY')
+        self.assertRaises(ValueError, vFrequency.from_ical, 234)
 
     def test_prop_vRecur(self):
         from icalendar.prop import vRecur